Fix split-package regression between orientdb-distributed and orientdb-server
